requirements.txt
streamlit
python-docx
python-dotenv
google-genai
pymupdf